added `operation_name` attribute to API G. method
This commit adds the `operation_name` attribute to the `aws_api_gateway_method`
resource. This attribute allows users to specify a custom function name to be
used when using API Gateway to generate an SDK. The attribute is already
supported by the AWS API and SDKs so there was really no reason not to include
it. It also made it a pretty straight forward change. This commit also includes:

* test cases for the new attribute
* updated documentation for the resource

Closes #13232