Use cfn change sets to determine update necessity
A reason for https://github.com/terraform-providers/terraform-provider-aws/issues/5194
is that after an completed rollback every successive update
with an "empty" change set will cause an error.
The logic is:
- initiate Update
- Wait for end state
- check for fail state

If there was no update performed on the stack the previous
fail state did not change and will also produce a fail state.
This causes that application of terraform plans will also fail
if they do not cause the stack to transitions to a success state. This usually involved a "dumb" update like introducing
or changing a tag and changing it back after a successful update.
The new implementation creates a change set and checks if the
list of changes is > 0. If there are no changes to be performed
the update function for the stack will be exited.