resource/aws_eks_node_group: Only pass ReleaseVersion during UpdateNodegroupVersion if changed
Reference: https://github.com/terraform-providers/terraform-provider-aws/issues/12675

It was expected to see an error similar to the issue report about the incompatible update and Terraform was previously submitting the state value for `release_version`, however the EKS API as of today was automatically fixing the incorrect value instead of returning an error. The resource change is to ensure that we will only submit correct API parameters should the API return errors for this situation in the future again.

Output from acceptance testing:

```
--- PASS: TestAccAWSEksNodeGroup_disappears (1356.93s)
--- PASS: TestAccAWSEksNodeGroup_InstanceTypes (1519.46s)
--- PASS: TestAccAWSEksNodeGroup_ScalingConfig_MaxSize (1540.56s)
--- PASS: TestAccAWSEksNodeGroup_DiskSize (1550.40s)
--- PASS: TestAccAWSEksNodeGroup_basic (1581.83s)
--- PASS: TestAccAWSEksNodeGroup_AmiType (1591.47s)
--- PASS: TestAccAWSEksNodeGroup_RemoteAccess_SourceSecurityGroupIds (1602.60s)
--- PASS: TestAccAWSEksNodeGroup_ScalingConfig_DesiredSize (1632.41s)
--- PASS: TestAccAWSEksNodeGroup_ScalingConfig_MinSize (1683.52s)
--- PASS: TestAccAWSEksNodeGroup_Tags (1712.30s)
--- PASS: TestAccAWSEksNodeGroup_Labels (1765.19s)
--- PASS: TestAccAWSEksNodeGroup_RemoteAccess_Ec2SshKey (1767.20s)
--- PASS: TestAccAWSEksNodeGroup_ReleaseVersion (2853.92s)
--- PASS: TestAccAWSEksNodeGroup_Version (3045.62s)
```