Only pass Workspace AutoStopTimeout for AutoStop
When specifying RunningMode as RunningModeAlwaysOn, aws-sdk-go throws an
error when the property RunningModeAutoStopTimeoutInMinutes is passed,
causing "terraform apply" to fail, passing up the error from aws-sdk-go:

"workspace creation failed: RunningModeAutoStopTimeoutInMinutes is not
applicable for WorkSpace with running mode set to ALWAYS_ON"

This happens even if you have not explicitly specified the timeout
property in Terraform config.

While aws-sdk-go could arguably ignore the irrelevant property, this
change simply omits it from the properties struct unless the running
mode is AutoStop.