New Guide: Resource Tagging
This guide presents a walkthrough of various Terraform AWS Provider functionality relating to resource tags. It especially highlights newer functionality such as provider level ignore tags configuration and individual tag resources. As with other guides originating in this repository, this may make sense to be migrated to the Learn website in the future.