docs/rds_cluster_instance: the engine/version should match that of the cluster
If you try to create an RDS instance with a different engine/version to
the cluster, you get an InvalidParameterCombination error from Terraform
and no instances.

Changing the behaviour so that instances inherit their config from the
cluster is a big change.

This patch just tweaks the documented examples, so if people copy/paste
the example and change the engine/version of the cluster, it will still
work correctly.  It also highlights that the engine/version should
match between the instance and the cluster.

Relates to https://github.com/terraform-providers/terraform-provider-aws/issues/10585