r/aws_route_table: Delete 'TestAccAWSRouteTable_IPv4_To_NetworkInterface_Attached' and 'TestAccAWSRouteTable_IPv4_To_NetworkInterface_TwoAttachments'.
If we mark `instance_id` and `network_interface_id` as both `Computed: true` in the `route` attribute's schema
then we end up having to change the associated set hash function to choose one or other of those attributes to
include in the hash and ignore the other. This means that either 'TestAccAWSRouteTable_IPv4_To_Instance' will
show continuous diffs or 'TestAccAWSRouteTable_IPv4_To_NetworkInterface_Attached' will.
The longer term solution is to remove `instance_id` as a route target as the instance's primary ENI's ID can
be used in the `network_interface_id`. This will also simplify the `aws_route` resource.