support spot price data by solely filtering
add a test for sourcing an aws_ec2_spot_price data source using
filter{} blocks exclusively. this commit also involves a rework of
the DescribeSpotPriceHistory AWS API call, namely using the native
SDK paging function; without this function, making a singular direct
call to the DescribeSpotPriceHistory SDK function returned a response with
an empty slice, but the underlying API call returned a response with a non-nil
NextToken value. to support this, we simply use the
DescribeSpotPriceHistoryPages unconditionally during reads, and
add all found SpotPriceHistory objects to a slice.