Add wait_until_stable option for ECS services
There are quite a few elements of an ECS service that cause the resource
to be recreated. Upon creation, we do not wait for the service to be
stable, which means even using lifecycle rules to
`create_before_destroy`, we will suffer downtime.

This commit seeks to introduce a basic assessment on whether a service
is stable. The attributes for this assessment are:

- if the desired count matches the running count
- if the pending count is zero
- if the service is described as "ACTIVE"

If all 3 of these conditions are met, we can be reasonably sure that the
service has at least started and provisioned some tasks. It won't
account for tasks that end up crashing, but if you're confident of your
service it should be safe to use.