New Data Source: aws_ec2_spot_price (#12504)
* d/aws_ec2_instance_spot_price - add new data source

Adds a data source to fetch the most recently published Spot Price
value for a given EC2 instance type and availability zone. The value
can be manipulated and fed into spot price parameters in other resources,
allowing for dynamic adjusted of spot requests.

* remove accidental inclusion of local test example

* make instance_type and availability_zone optional

* simplify DescribeSpotPriceHistory usage

since we are only interested in fetching the latest SpotPrice,
we don't need to iterate over a series of requests, as we only
ever need to fetch call's worth of data

* copypasta

* add more thorough tests for aws_ec2_instance_spot_price resource

* use dynamic data sources for filtering in tests

* format spot_price_timestamp as RFC3339

* rebase aws/provider.go

* fix up tests

* lint and typo fixes

* update docs

* whitespace

* aws_ec2_instance_spot_price -> aws_ec2_spot_price

* support spot price data by solely filtering

add a test for sourcing an aws_ec2_spot_price data source using
filter{} blocks exclusively. this commit also involves a rework of
the DescribeSpotPriceHistory AWS API call, namely using the native
SDK paging function; without this function, making a singular direct
call to the DescribeSpotPriceHistory SDK function returned a response with
an empty slice, but the underlying API call returned a response with a non-nil
NextToken value. to support this, we simply use the
DescribeSpotPriceHistoryPages unconditionally during reads, and
add all found SpotPriceHistory objects to a slice.

* update documentation following resource rename

* copypasta documention updates