Add cloudwatch_event_rules data source
This is a data source used to query event rules by name_prefix, and
returns a list of rules names.

The motivation for this data source is that I have a deployment that
create event rules outside of Terraform, but I wish to be able to create
different resources based upon the event rules that have been created.

I also intend to add a cloudwatch_event_rule data source that provides
full details for an event rule object.