Add conflicts between cidr_blocks and self in docs
When reading the docs, it suggested that you could use `cidr_blocks` and `self` together. I tried that and got this error:

```
Error: 'self': conflicts with 'cidr_blocks' ([]interface {}{"10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16"})
```

This change updates the docs to mention that cidr_blocks and self are not allowed to be used with each other.