resource/aws_s3_bucket_object: Ignore changes due to default bucket encryption (#10200)
When an object is uploaded to a bucket that has default encryption configured, the object inherits the kms_key_id from the bucket. We therefore need to ignore changes to that attribute, which the user won't have specified. Setting this attribute to "" has the effect of re-encrypting the object with the default `aws/s3` key, which is not what the user intended or expected.