Doc: `VPC` for endpoint_type to transfer_server
Document `VPC` value for endpoint_type attribute in transfer_server resource.

I put `VPC_ENDPOINT` in parentheses because it is no longer possible to create this type of Transfer servers in AWS dashboard. It is only accessible via the API, so I believe it will be eventually deprecated, but I didn't find any information to support that.