provider: Update Makefile fmt target to ignore ./awsproviderlint/vendor
Reference: https://github.com/terraform-providers/terraform-provider-aws/pull/15391

Previously was not an issue before because we were vendoring in the root directory and explicitly referencing package subdirectories. Now that a vendor directory is included in ./awsproviderlint/ we need to exclude it. Using Makefile's built-in filter-out and wildcard functions to avoid platform specific issues with find, grep, etc.