resource/aws_elastic_beanstalk_environment: Prevent spurious settings updates for custom load balancer listener rules
If you specified a load balancer rule like this in your Elastic
Beanstalk environment settings:
setting {
  namespace = "aws:elbv2:listenerrule:SharedAlbRedirect"
  name      = "HostHeaders"
  value     = "www.myapp.com"
}
setting {
  namespace = "aws:elbv2:listener:443"
  name      = "Rules"
  value     = "SharedAlbRedirect"
}

Terraform would continuously detect environment changes on apply,
showing plan diff like:

+ setting {
    + name      = "HostHeaders"
    + namespace = "aws:elbv2:listenerrule:SharedAlbRedirect"
    + value     = "www.myapp.com"
  }
- setting {
    - name      = "HostHeaders" -> null
    - namespace = "aws:elbv2:listenerrule:SharedAlbRedirect" -> null
    - value     = "www.myapp.com,abcdef123.us-east-1.elasticbeanstalk.com" -> null
  }

Because Elastic Beanstalk automatically adds the environment's CNAME to
your load balancer rules.

It's not possible for users to add this record to their rules, as it's
not predictable before the environment is created.[1] So, we hide it
from the user, similar to how we hide EB's generated security group in
the "SecurityGroups" setting's value.

[1] Well, it is mostly predictable if you specify a cname_prefix. But
that's (IMO) an uncommon option.