resource/aws_db_instance: Ignore DBInstanceNotFound error during deletion (#15942)
Reference: https://github.com/terraform-providers/terraform-provider-aws/issues/12900

Unfortunately testing this scenario is complicated in the Terraform Plugin SDK acceptance testing framework as refresh before deletion is standard behavior and the error gets caught in the Read function (e.g. disappears testing). In the future, we may propose to write a standard way for converting AWS Go SDK errors with specific error codes and messages to the Terraform Plugin SDK resource.NotFoundError, which we could then use static analysis against in the Read/Delete/CheckDestroy functions to ensure its always handled. For now though, this is a targeted fix for a community reported situation.