infrastructure/repository: Bump Terraform CLI to 0.13.5 and use remote backend
The `atlas` backend is legacy so migrated to `remote`. Upgraded the Terraform CLI version in Terraform Cloud and verified locally with TFC token:

```console
$ terraform plan
...
No changes. Infrastructure is up-to-date.

This means that Terraform did not detect any differences between your
configuration and real physical resources that exist. As a result, no
actions need to be performed.
```