resource/aws_emr_cluster: fix error on missing cluster
After a certain amount of time, destroyed EMR clusters completely
disappear from AWS API output. They don't just display some deleted
state, they're just gone.

Currently, if you have one of these clusters in a statefile and you
never update it in between the time when it was deleted and when it is
purged from the API, terraform crashes the next time you try to run it.
This patch fixes that.