Add default values to aws_vpn_connection
This fixes a bug where removing fields from the resource should revert
them to their default value, but instead sets them to their empty value
(in this case 0). When this happens the provider attempts to use these 0
values to update the VPN connection which results in an error, forcing
the user to specify these fields and their default values.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>