.github/workflows: Reduce goreleaser parallelism to match GitHub Actions hosted runners (#17067)
Most of the GitHub Actions logging is either missing or cutoff without explanation, but finally found one with a `signal: killed` entry. We may be triggering oom-killer or another abuse protection mechanism, so this attempts to reduce the footprint on the runner. I did not immediately see a way to dynamically fetch the processor count via workflow information.