aws_instance: fix for setting iops+throughput at the same time
When both `iops` and `throughput` are specified on an inline block
device (`root_block_device` or `ebs_block_device`), both attributes
should be honored. With the `else if`, specifying `iops` would make the
`throughput` attribute a NOOP during instance creation.

The desired throughput shows up correctly in the plan, but after apply,
it is actually the default (125) in the state and in EC2.

Signed-off-by: Matthias Rampke <matthias@rampke.de>