Greetings.  This is my first attempt to contribute to terraform and go is not even slightly my first language.
This is to resolve 12373 which I would like to resolve for selfish reasons as it affects me.  It seemed like a good place to start.

The new data source "aws_ec2_transit_gateway_route_tables" is modelled after "aws_route_tables".  The code is shamelessly plagiarized from the aws_route_tables and the aws_transit_gateway_route_table resources.

I've included tests which I think are right, but I'm not sure on protocol for running them.

The code compiles.  Which is a good start.