r/aws_mq_configuration: add missing authentication_strategy
When creating an ActiveMQ instance with authentication_strategy=ldap,
then the corresponding configuration has to contain the same option
as well for the user to be able to include a cachedLDAPAuthorizationMap
element in the broker xml configuration. If the option is not provided
when creating the configuration, the default strategy "simple" is
assumed and AWS automatically removes the cachedLDAPAuthorizationMap
element.