service/storagegateway: Support Cached and VTL gateway upload buffers
Reference: https://github.com/hashicorp/terraform-provider-aws/issues/17809

The referenced issue contains the context, but the gist of this issue is that the Storage Gateway API canonicalizes the Cached and VTL disk identifiers only after first cache or upload buffer use, which means that this resource must accept the path first, then perform its own lookup after creation.

This also fixes the `aws_storagegateway_local_disk` data source, which was missing `Computed` on the `disk_node` and `disk_path` attributes, which prevented it for being used to lookup one for the value of the other.

Previously:

```
=== CONT  TestAccAWSStorageGatewayUploadBuffer_DiskPath
    resource_aws_storagegateway_upload_buffer_test.go:107: Step 1/2 error: Error running apply: exit status 1

        Error: error reading Storage Gateway Upload Buffer (arn:aws:storagegateway:us-west-2:187416307283:gateway/sgw-D0A941B9:/dev/nvme1n1): not found

          on terraform_plugin_test.tf line 129, in resource "aws_storagegateway_upload_buffer" "test":
         129: resource "aws_storagegateway_upload_buffer" "test" {

--- FAIL: TestAccAWSStorageGatewayUploadBuffer_DiskPath (418.35s)
```

Output from acceptance testing:

```
--- PASS: TestAccAWSStorageGatewayLocalDiskDataSource_DiskNode (288.52s)
--- PASS: TestAccAWSStorageGatewayLocalDiskDataSource_DiskPath (300.60s)

--- PASS: TestAccAWSStorageGatewayUploadBuffer_basic (418.26s)
--- PASS: TestAccAWSStorageGatewayUploadBuffer_DiskPath (444.33s)
```