tests/provider: Fix and enable tfproviderlint AT012 for mismatched acceptance test prefixes within a single file
Reference: https://github.com/hashicorp/terraform-provider-aws/issues/18377

This selection process was just choosing the most prominent naming convention within the files, not making any assertions about which is better.

The ELB "Update" tests were just configuration update versions of the non-"Update" tests so consolidated to just the updating tests. The EB import test was the same as the basic test so moved the import step to the basic test as is customary.

Output from acceptance testing for modified tests:

```
--- PASS: TestAccAWSBeanstalkApp_basic (17.01s)
```