Clarify key policy vs IAM policy
Key policies are not the same as IAM policies. While key policies are technically required for all KMS keys, KMS keys that are created programmatically are given a default policy that allows the `:root` of an account full access to the key, which allows IAM policies to be applied to the key.