fix(r_aws_mwaa_environment): Changing the KMS key used by MWAA should recreation of the Airflow environment
This is currently an issue when updating the cluster from `aws/airflow` KMS key to a CMK for example. The current implementation will show a diff, apply an update to the cluster but will not be able to change to another KMS key leading to a perpetual diff.

```
Terraform will perform the following actions:

  # module.airflow[0].aws_mwaa_environment.airflow will be updated in-place
  ~ resource "aws_mwaa_environment" "airflow" {
        id                              = "myairflow"
      + kms_key                         = "arn:aws:kms:us-east-1:redacted:key/redacted"
        name                            = "myairflow"
        tags                            = {
            "Name"                 = "myairflow"
        }
        # (21 unchanged attributes hidden)

        # (2 unchanged blocks hidden)
    }

Plan: 0 to add, 1 to change, 0 to destroy.
```