Update rds_global_cluster.html.markdown
Update the MySQL Global Cluster example to work around an issue where the engine parameters for the individual cluster instances are not inherited from the cluster and must be explicitly defined again. Without this fix the published example would not build with AWS Provider 3.49 and engine 5.7.mysql_aurora.2.10.0, resulting in:

Error: error creating RDS Cluster (my-cluster-pri) Instance: InvalidParameterCombination: Cannot find version 5.7.mysql_aurora.2.10.0 for aurora

The cluster in the other region will then also fail with:
Error: error creating RDS cluster: InvalidParameterValue: The parameter MasterUsername must be provided and must not be blank.

These two separate errors from the separate clusters then confound troubleshooting because it's not clear which is cause and which is effect. The second error is in fact spurious and only occurs because the primary cluster failed to build.