Allow lazy AWSClient initialization.
AWSClient is currently statically initialized in `Config.Client()`,
called by `providerConfigure`.
This makes using the provider in a configuration-driven template
dependant on having a valid set of aws credentials, even if the
configuration provides for no aws resources to be managed.

This commit, while keeping the same approach as a default, allows the
user to configure this initialization to be delayed until (and if)
necessary.

In more detail, this commit/PR:
- adds the provider configuration to `AWSClient.providerConfig`
- moves the `AWSClient` initialization process to a new
  `AWSClient.Init() error` function.
- adds a new provider parameter `lazy_init` (default `false`)
- calls `AWSClient.Init` from `Config.Client()` only if `lazy_init`
  if `true`
- adds a series of
  `ClientInit{CrudBase,CrudContext,Exists,MigrateState,CustomizeDiff,StateUpgrade}Func`
  wrapper functions used in each resource definition to integrate
  the current `Provider` arg function functionalities with a call to
  `AWSClient.Init()`.