Update elasticsearch version & snapshot details
Few people would still be using such an old version of ElasticSearch, so I updated the ElasticSearch version of the example to the latest. Related to this is that the snapshot options are deprecated since 5.3 (and now we're on 7.10). Removing the snapshot options from the example will help those like me who copy/paste from Terraform documentation examples.