Fix import of aws_route53_record with underscore
If a record contained an underscore in the record name but the
underscore was not the first character of the record, the record failed
to import.

This change resolves the issue by using the type as a known point for
splitting the record. As long as there is a valid record type in the
parts of the ID, we can determine where the type and set identifier
begin, and the record name is the rest.