Add timeout support for EBS Snapshot copies
These copies can take an exorbitant amount of time to run, and if a
snapshot requires a custom timeout configuration, chances are
extremely high that any attempt to copy that snapshot will also exceed
the default 10-minute timeout.

Thankfully, the API is very similar between the base snapshot resource
and the copy resource, so we can borrow the logic almost wholesale
from the EBS snapshot to support this configuration block
appropriately.

Resolves #14354.