Add eventual consistency protection to efs_file_system_policy
There was no retry logic around putting an EFS file system policy, which
would lead to errors when the same TF both creates a new IAM principal
(e.g., role) and tries to specify that principal in an EFS FS policy in
the same run. This adds retry logic when AWS returns an error about an
invalid principal in line with standard recommendations.

Fixes #21716