Fix for hashicorp/terraform-provider-aws/#20452
When using `wait_for_stead_state` retry up to 3 times.

This is due to when a service is replaced there is a possibility the
service will return a status of INACTIVE as the AWS API returns the
status of the old service instead of the new one which hasn't fully
registered yet.