Add note about filename
This recently bit me. Took me a bit to figure it out. Maybe I'm dumb ¯\\_(ツ)_/¯

I had a lambda resource in a module. Based on the sample I thought I had done everything right. Then I realized that because this was a module I needed to specify more than just the filename. 

I think where I get screwed up is that filename, to me, means just the name of the file but in reality this is path and filename.

I also didn't see this the first time around:

```
* `filename` - (Optional) Path to the function's deployment package within the local filesystem. Conflicts with `image_uri`, `s3_bucket`, `s3_key`, and `s3_object_version`.
```

Still feels really nuanced? Feel free to close if this is just me not RTFM'ing enough.