examples: configuration for intra-region Transit Gateway Peering
Issue #23828 made me think that the intra-region Peering use-case
is not straight forward. Operators might assume they need to
configure the Acceptor to the Peering Attachment that Terraform creates.
In fact, AWS creates two Peering Attachment resources, one for the
Creator side and one for the Acceptor side. So, the a Data Source
is needed to find the second Acceptor side's attachment and then
use it to configure the Acceptor resource.

Closes: #23828