Treat Deployment Targets as their own case
The original code tried to add deployment target support by finding
an account in one of the targets, and then using all the same code
as account stack set instances. This is deficient:

- If none of the targets contain an account the code fails, due to
  the resource Id not having an accountID value.
- It does not expose information like account IDs and stack set IDs
  for any accounts in the deployment targets, except the arbitrary
  accoun it used.
- It can cause the resource Id to needlessly drift due to the
  account selection process changing (account moved OU's, or the
  API decided to return things in a different order, etc).

This refactor treats deployment targets as their own branch of the
logic, with their own new attributes.