service/iam: iam_principal_policy_simulation data source
This data source wraps the IAM policy simulation API. This was previously
a data source with little utility in Terraform, but with the introduction
of preconditions and postconditions in Terraform v1.2.3 it can be
potentially useful as a way for a configuration to either pre-verify that
it seems to be running with credentials that confer sufficient access or
to self-check a policy it declares itself to get earlier warning if the
policy seems insufficient for its intended purpose.

Unfortunately the IAM policy simulator is pretty low-level and requires
the caller to figure out all of the relevant details of how a real AWS
service would make requests to IAM at runtime in order to construct
a fully-realistic simulation, but thankfully in practice it seems like
authors could make do with relatively-simple "naive" simulations unless
they know they are using more complex IAM policy features, such as custom
conditions or interpolations.