Example isn't automation friendly
The example requires the lambda to be zipped before terraform is run, which means it doesn't work when copy and pasted. 

I updated the example to be automation and CI/CD friendly, and move the zipping into the terraform code using the archive_file data source. This works with copy and pasting.