Correct documentation to specify ARN is required. (#28661)
When following the documentation, the following error is returned:

```
Error: "rule.0.destination.0.bucket" (given-bucket-name) is an invalid ARN: arn: invalid prefix
```

This is because the API requires the full ARN be given. The resource
correctly identifies `verify.ValidARN`, but the documentation is
incorrect. This is updated in accordance with AWS's own documentation
that can be referenced in [the official API documentation](https://docs.aws.amazon.com/AmazonS3/latest/API/API_Destination.html).