Ensure required values are set when changing `storage_type`.
As observed in #28589, if changing the `storage_type` additional fields
need to be set if required.

For example, `gp3` storage requires that the `allocated_storage` and `iops`
params are supplied if the volume size is `>=400G`.

Added a new test case to try and exercise this behaviour, and also tweaked
a couple of existing tests to exercise the `gp3` tiering behaviour.
**N.B** I've not been able to run these acceptance tests locally yet.

Fixes #28589