feature(lambda): Support CRUD
Add a lifecycle_scope argument which allows configuring the lambda_invocation
resource to trigger for each lifecycle transition of the terraform resource.
By adding the argument it is possible to maintain backwards compatibility
and enable lambda invocations for resource updates/deletes.

See the updated documentation for a detailed explanation of the feature.