Show safer example for `device_name`, and mark it as required
`device_name` is required in a `block_device_mappings` block. `terraform
validate` doesn't complain when it's missing but `apply` will come back
with a HTTP 400 error.

Secondly, the top example shows using `/dev/sda1` for an additional
volume. Doing this can cause you to end up with unbootable instances,
even if the root volume of the AMI is /dev/xvda. Moreover, the web
management console won't allow you to select `/dev/sda` when setting up
a launch template.

According to
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/device_naming.html,
`/dev/sdf` is a safe choice for both HVM and Paravirtual instances.