dms: Add db2-zos engine type
Looks like this released in July 2022:
https://aws.amazon.com/about-aws/whats-new/2022/07/aws-database-migration-ibm-db2-zos-source/

From looking at the docs:
https://docs.aws.amazon.com/sdk-for-go/api/service/databasemigrationservice/
it seems that the same IBMDb2Settings is used for both `db2` and `db2-zos`.

For tests, I duplicated the existing db2 tests and ensured the new
`db2-zos` engine_name was used. Given how non-unique the code changes
are, these tests might be redundant.