fix: ensure endpoint group weight 0 is taken into account
Zero is a valid value for an endpoint group configuration and
it shouldn't be ignored and not sent to the API.

The current behaviour causes a endpoint weight of 128 to be used
instead as it is AWS default value if no weight is set.
This causes traffic to be sent despite the expectations that
a zero value will cause no traffic on the endpoint.