Update organization data source
Check if current account is a delegated administrator for the
Organization. If so, it can also read the roots, account list, etc.

If the ListDelegatedAdministrators call returns Access Denied, ignore
the error as it means the account is just a regular child account. If it
wasn't in the Organization at all, the previous DescribeOrganization
call would have failed anyway.