ci: fix diff grep regex
Adjusts the grep regular expression so that only added import lines will match.
Previously unmodified lines which are displayed with the diff for context
could also match the provided expression. Now the line must start with a
literal +, contain only whitespace prior to a double quote, and match
the root path of the AWS SDK for Go v1 library.