r/aws_elasticache_replication_group: v2 state upgrader
This state upgrader resolves issues with the default auth_token_update_strategy value, introduced in v5.27.0, which causes a diff upon upgrade and triggers an attempted in-place update.
The default value will now instead be written to state by the state upgrader, resulting in no changes on the initial apply.