r/aws_elasticache_replication_group(doc): clarify auth_token_update_strategy argument dependencies
While this value is only sent to the AWS API if auth_token is set (and by association, transit_encryption_enabled is
set to true), the default value means it is always set in state. As such, indiciating in the documentation that
it can only specified when transit encrpytion is enabled is no longer true.