Mention "cloudfront-js-2.0" now is an available runtime for CloudFront funcs
This seems to have been a recent addition that hasn't been announced anywhere.
It was mentioned in
https://aws.amazon.com/blogs/aws/introducing-amazon-cloudfront-keyvaluestore-a-low-latency-datastore-for-cloudfront-functions/
but otherwise it just seems to have appeared in the documentation:

https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/functions-javascript-runtime-20.html
https://docs.aws.amazon.com/cloudfront/latest/APIReference/API_FunctionConfig.html

I have updated the runtime version shown in the example as I assume most people
would prefer to use the more recent runtime version.