fix: set redis engine version only uses .x suffix on version 6
When setting the redis engine version from an outdated terraform state it should only set the version as 6.x, but never as [7-9].x.
This changes the handling to consider an existing terraform state value for the engine version of 6.x only if the engine version returned from the Elasticache API is also of major version 6. Otherwise, the .x suffix is ignored and the engine version is set as <major>.<minor>